fixed issue #2086: Deleting restored objects produces error message in the left frame of the administration view
git-svn-id: http://svn.ilias.de/svn/ilias/trunk@13850 21b2c9ec-7c21-0410-8b45-9bfb8ed2bfc5

+++ Services/COPage/classes/class.ilPageObject.php
@@ -1254,10 +1254,14 @@ class ilPageObject
 			" AND parent_type= ".$ilDB->quote($this->getParentType());
 		$this->ilias->db->query($query);
 
+		// delete media objects
 		foreach ($mobs as $mob_id)
 		{
-			$mob_obj =& new ilObjMediaObject($mob_id);
-			$mob_obj->delete();
+			if (ilObject::_exists($mob_id))
+			{
+				$mob_obj =& new ilObjMediaObject($mob_id);
+				$mob_obj->delete();
+			}
 		}
 
 		include_once("./Modules/File/classes/class.ilObjFile.php");
